scsi: lpfc: Remove REG_LOGIN check requirement to issue an
From e78c006f4c888231cdabb8a4286ba17980a903fa Mon Sep 17 00:00:00 2001
 ELS RDF
Git-commit: e78c006f4c888231cdabb8a4286ba17980a903fa
Patch-mainline: v5.15-rc1
References: bsn#61

Since the REG_LOGIN to the fabric controller happens in parallel with SCR,
it may or may not be completed by the time RDF is sent.  RDF and SCR are
sent to the fabric in parallel, so checking for a completed REG_LOGIN in
the RDF submit path is not needed.

Remove the REG_LOGI check from the RDF submission path.

Link: https://lore.kernel.org/r/20210707184351.67872-11-jsmart2021@gmail.com
Co-developed-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.15-rc1-scsi-lpfc-Remove-REG_LOGIN-check-requirement-to-issu.patch