xfs: reset child dir '..' entry when unlinking child
From 5838d0356bb3c320867c393f12b169c01a870bda Mon Sep 17 00:00:00 2001
Git-commit: 5838d0356bb3c320867c393f12b169c01a870bda
Patch-mainline: v5.14-rc2
References: bsn#12

While running xfs/168, I noticed a second source of post-shrink
corruption errors causing shutdowns.

Let's say that directory B has a low inode number and is a child of
directory A, which has a high number.  If B is empty but open, and
unlinked from A, B's dotdot link continues to point to A.  If A is then
unlinked and the filesystem shrunk so that A is no longer a valid inode,
a subsequent AIL push of B will trip the inode verifiers because the
dotdot entry points outside of the filesystem.

To avoid this problem, reset B's dotdot entry to the root directory when
unlinking directories, since the root directory cannot be removed.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.14-rc2-xfs-reset-child-dir-.-entry-when-unlinking-child.patch